requirements.txt
pandas
google-cloud-bigquery
loguru
apache-airflow-providers-google
apache-airflow-providers-postgres
google-cloud-storage
